app/requirements.txt
gunicorn==20.1.0
Django==4.2
celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
dj-rest-auth[with_social]==4.0.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-simple-history==3.3.0
facebook-business==16.0.2
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
pytest==7.3.1
pytest-django==4.5.2
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
stripe==5.4.0
pydantic==1.10.8
openai==0.27.7
django-simple-history==3.3.0
